only one config-split-names step
