+++ libs/libcommon/src/libcommon/constants.py
@@ -51,7 +51,7 @@ EXTERNAL_DATASET_SCRIPT_PATTERN = "datasets_modules/datasets"
 CONFIG_HAS_VIEWER_KIND = "config-size"
 CONFIG_INFO_KIND = "config-info"
 CONFIG_PARQUET_METADATA_KIND = "config-parquet-metadata"
-CONFIG_SPLIT_NAMES_KIND = "config-split-names-from-streaming"
+CONFIG_SPLIT_NAMES_KIND = "config-split-names"
 DATASET_CONFIG_NAMES_KIND = "dataset-config-names"
 DATASET_INFO_KIND = "dataset-info"
 SPLIT_DUCKDB_INDEX_KIND = "split-duckdb-index"
