+++ services/worker/src/worker/job_runners/config/split_names.py
@@ -12,8 +12,10 @@ from libcommon.exceptions import (
     DatasetManualDownloadError,
     DatasetWithScriptNotSupportedError,
     EmptyDatasetError,
+    PreviousStepFormatError,
     SplitNamesFromStreamingError,
 )
+from libcommon.simple_cache import CachedArtifactError, CachedArtifactNotFoundError, get_previous_step_or_raise
 
 from worker.dtos import CompleteJobResult, SplitsList
 from worker.job_runners.config.config_job_runner import ConfigJobRunnerWithDatasetsCache
@@ -27,13 +29,10 @@ def compute_split_names_from_streaming_response(
     hf_token: Optional[str] = None,
 ) -> SplitsList:
     """
-    Get the response of 'config-split-names-from-streaming' for one specific dataset and config on huggingface.co.
+    Get the response of 'config-split-names' for one specific dataset and config on huggingface.co, using streaming.
 
     This function relies on the streaming mode if the splits are not directly defined in the dataset config. See
     https://github.dev/huggingface/datasets/blob/e183a269067575db8765ee979bd8523d14a1adae/src/datasets/inspect.py#L389-L390
-
-    The config-split-names-from-streaming response generated by this function does not include stats about the split,
-    like the size or number of samples. See dataset-info or dataset-size for that.
 
     Args:
         dataset (`str`):
@@ -62,7 +61,7 @@ def compute_split_names_from_streaming_response(
     Returns:
         `SplitsList`: An object with the list of split names for the dataset and config.
     """
-    logging.info(f"compute 'config-split-names-from-streaming' for {dataset=} {config=}")
+    logging.info(f"compute 'config-split-names' using streaming for {dataset=} {config=}")
     try:
         split_name_items: list[FullSplitItem] = [
             {"dataset": dataset, "config": config, "split": str(split)}
@@ -91,17 +90,62 @@ def compute_split_names_from_streaming_response(
     return SplitsList(splits=split_name_items)
 
 
-class ConfigSplitNamesFromStreamingJobRunner(ConfigJobRunnerWithDatasetsCache):
+def compute_split_names_from_info_response(dataset: str, config: str) -> SplitsList:
+    """
+    Get the response of 'config-split-names' for one specific dataset and config on huggingface.co
+    computed from cached response in dataset-info step.
+
+    Args:
+        dataset (`str`):
+            A namespace (user or an organization) and a repo name separated
+            by a `/`.
+        config (`str`):
+            A configuration name.
+
+    Raises:
+        [~`libcommon.simple_cache.CachedArtifactError`]:
+          If the previous step gave an error.
+        [~`libcommon.simple_cache.CachedArtifactNotFoundError`]:
+          If the previous step has not been computed yet.
+        [~`libcommon.exceptions.PreviousStepFormatError`]:
+          If the content of the previous step has not the expected format
+
+    Returns:
+        `SplitsList`: An object with the list of split names for the dataset and config.
+    """
+    logging.info(f"compute 'config-split-names' from config-info for {dataset=} {config=}")
+    config_info_response = get_previous_step_or_raise(kind="config-info", dataset=dataset, config=config)
+
+    try:
+        splits_content = config_info_response["content"]["dataset_info"]["splits"]
+    except Exception as e:
+        raise PreviousStepFormatError("Previous step 'config-info' did not return the expected content.") from e
+
+    split_name_items: list[FullSplitItem] = [
+        {"dataset": dataset, "config": config, "split": str(split)} for split in splits_content
+    ]
+
+    return SplitsList(splits=split_name_items)
+
+
+class ConfigSplitNamesJobRunner(ConfigJobRunnerWithDatasetsCache):
     @staticmethod
     def get_job_type() -> str:
-        return "config-split-names-from-streaming"
+        return "config-split-names"
 
     def compute(self) -> CompleteJobResult:
-        return CompleteJobResult(
-            compute_split_names_from_streaming_response(
-                dataset=self.dataset,
-                config=self.config,
-                hf_token=self.app_config.common.hf_token,
-                dataset_scripts_allow_list=self.app_config.common.dataset_scripts_allow_list,
+        try:
+            return CompleteJobResult(compute_split_names_from_info_response(dataset=self.dataset, config=self.config))
+        except (CachedArtifactError, CachedArtifactNotFoundError):
+            logging.info(
+                f"Cannot compute 'config-split-names' from config-info for {self.dataset=} {self.config=}. "
+                f"Trying to compute it using streaming."
             )
-        )
+            return CompleteJobResult(
+                compute_split_names_from_streaming_response(
+                    dataset=self.dataset,
+                    config=self.config,
+                    hf_token=self.app_config.common.hf_token,
+                    dataset_scripts_allow_list=self.app_config.common.dataset_scripts_allow_list,
+                )
+            )
