--- services/worker/src/worker/job_runners/config/split_names_from_info.py
@@ -1,59 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# Copyright 2022 The HuggingFace Authors.
-
-import logging
-
-from libcommon.dtos import FullSplitItem
-from libcommon.exceptions import PreviousStepFormatError
-from libcommon.simple_cache import get_previous_step_or_raise
-
-from worker.dtos import CompleteJobResult, SplitsList
-from worker.job_runners.config.config_job_runner import ConfigJobRunner
-
-
-def compute_split_names_from_info_response(dataset: str, config: str) -> SplitsList:
-    """
-    Get the response of 'config-split-names-from-info' for one specific dataset and config on huggingface.co
-    computed from cached response in dataset-info step.
-
-    The 'config-split-names-from-info' response generated by this function does not include stats about the split,
-    like the size or number of samples. See dataset-info or dataset-size for that.
-
-    Args:
-        dataset (`str`):
-            A namespace (user or an organization) and a repo name separated
-            by a `/`.
-        config (`str`):
-            A configuration name.
-
-    Raises:
-        [~`libcommon.simple_cache.CachedArtifactError`]:
-          If the previous step gave an error.
-        [~`libcommon.exceptions.PreviousStepFormatError`]:
-          If the content of the previous step has not the expected format
-
-    Returns:
-        `SplitsList`: An object with the list of split names for the dataset and config.
-    """
-    logging.info(f"compute 'config-split-names-from-info' for {dataset=} {config=}")
-    config_info_response = get_previous_step_or_raise(kind="config-info", dataset=dataset, config=config)
-
-    try:
-        splits_content = config_info_response.response["content"]["dataset_info"]["splits"]
-    except Exception as e:
-        raise PreviousStepFormatError("Previous step 'config-info' did not return the expected content.") from e
-
-    split_name_items: list[FullSplitItem] = [
-        {"dataset": dataset, "config": config, "split": str(split)} for split in splits_content
-    ]
-
-    return SplitsList(splits=split_name_items)
-
-
-class ConfigSplitNamesFromInfoJobRunner(ConfigJobRunner):
-    @staticmethod
-    def get_job_type() -> str:
-        return "config-split-names-from-info"
-
-    def compute(self) -> CompleteJobResult:
-        return CompleteJobResult(compute_split_names_from_info_response(dataset=self.dataset, config=self.config))
