+++ services/worker/tests/job_runners/config/test_split_names.py
@@ -4,16 +4,21 @@
 from collections.abc import Callable
 from dataclasses import replace
 from http import HTTPStatus
+from typing import Any
 
 import pytest
 from libcommon.dtos import Priority
-from libcommon.exceptions import CustomError, DatasetManualDownloadError
+from libcommon.exceptions import CustomError, DatasetManualDownloadError, PreviousStepFormatError
 from libcommon.resources import CacheMongoResource, QueueMongoResource
-from libcommon.simple_cache import upsert_response
+from libcommon.simple_cache import (
+    CachedArtifactError,
+    CachedArtifactNotFoundError,
+    upsert_response,
+)
 
 from worker.config import AppConfig
-from worker.job_runners.config.split_names_from_streaming import (
-    ConfigSplitNamesFromStreamingJobRunner,
+from worker.job_runners.config.split_names import (
+    ConfigSplitNamesJobRunner,
     compute_split_names_from_streaming_response,
 )
 from worker.resources import LibrariesResource
@@ -21,7 +26,7 @@ from worker.resources import LibrariesResource
 from ...fixtures.hub import HubDatasetTest, get_default_config_split
 from ..utils import REVISION_NAME
 
-GetJobRunner = Callable[[str, str, AppConfig], ConfigSplitNamesFromStreamingJobRunner]
+GetJobRunner = Callable[[str, str, AppConfig], ConfigSplitNamesJobRunner]
 
 
 @pytest.fixture
@@ -34,7 +39,7 @@ def get_job_runner(
         dataset: str,
         config: str,
         app_config: AppConfig,
-    ) -> ConfigSplitNamesFromStreamingJobRunner:
+    ) -> ConfigSplitNamesJobRunner:
         upsert_response(
             kind="dataset-config-names",
             dataset=dataset,
@@ -43,9 +48,9 @@ def get_job_runner(
             http_status=HTTPStatus.OK,
         )
 
-        return ConfigSplitNamesFromStreamingJobRunner(
+        return ConfigSplitNamesJobRunner(
             job_info={
-                "type": ConfigSplitNamesFromStreamingJobRunner.get_job_type(),
+                "type": ConfigSplitNamesJobRunner.get_job_type(),
                 "params": {
                     "dataset": dataset,
                     "revision": REVISION_NAME,
@@ -63,7 +68,97 @@ def get_job_runner(
     return _get_job_runner
 
 
-def test_compute(app_config: AppConfig, get_job_runner: GetJobRunner, hub_public_csv: str) -> None:
+@pytest.mark.parametrize(
+    "dataset,upstream_status,upstream_content,error_code,content",
+    [
+        (
+            "ok",
+            HTTPStatus.OK,
+            {
+                "dataset_info": {
+                    "splits": {
+                        "train": {"name": "train", "dataset_name": "ok"},
+                        "validation": {"name": "validation", "dataset_name": "ok"},
+                        "test": {"name": "test", "dataset_name": "ok"},
+                    },
+                }
+            },
+            None,
+            {
+                "splits": [
+                    {"dataset": "ok", "config": "config_name", "split": "train"},
+                    {"dataset": "ok", "config": "config_name", "split": "validation"},
+                    {"dataset": "ok", "config": "config_name", "split": "test"},
+                ]
+            },
+        ),
+        (
+            "upstream_fail",
+            HTTPStatus.INTERNAL_SERVER_ERROR,
+            {"error": "error"},
+            CachedArtifactError.__name__,
+            None,
+        ),
+        (
+            "without_dataset_info",
+            HTTPStatus.OK,
+            {"some_column": "wrong_format"},
+            PreviousStepFormatError.__name__,
+            None,
+        ),
+        (
+            "without_config_name",
+            HTTPStatus.OK,
+            {"dataset_info": "wrong_format"},
+            PreviousStepFormatError.__name__,
+            None,
+        ),
+        (
+            "without_splits",
+            HTTPStatus.OK,
+            {"dataset_info": {"config_name": "wrong_format"}},
+            PreviousStepFormatError.__name__,
+            None,
+        ),
+    ],
+)
+def test_compute_from_info(
+    app_config: AppConfig,
+    get_job_runner: GetJobRunner,
+    dataset: str,
+    upstream_status: HTTPStatus,
+    upstream_content: Any,
+    error_code: str,
+    content: Any,
+) -> None:
+    config = "config_name"
+    upsert_response(
+        kind="config-info",
+        dataset=dataset,
+        dataset_git_revision=REVISION_NAME,
+        config=config,
+        content=upstream_content,
+        http_status=upstream_status,
+    )
+    job_runner = get_job_runner(dataset, config, app_config)
+
+    if error_code:
+        with pytest.raises(Exception) as e:
+            job_runner.compute()
+        assert e.typename == error_code
+    else:
+        assert job_runner.compute().content == content
+
+
+def test_doesnotexist(app_config: AppConfig, get_job_runner: GetJobRunner) -> None:
+    dataset = "non_existent"
+    config = "non_existent"
+    worker = get_job_runner(dataset, config, app_config)
+    with pytest.raises(CachedArtifactNotFoundError):
+        worker.compute()
+
+
+def test_compute_from_streaming(app_config: AppConfig, get_job_runner: GetJobRunner, hub_public_csv: str) -> None:
     dataset = hub_public_csv
     config, _ = get_default_config_split()
     job_runner = get_job_runner(dataset, config, app_config)
