+++ services/worker/tests/job_runners/split/test_descriptive_statistics.py
@@ -99,7 +99,7 @@ def get_job_runner(
         )
 
         upsert_response(
-            kind="config-split-names-from-streaming",
+            kind="config-split-names",
             dataset=dataset,
             dataset_git_revision=REVISION_NAME,
             config=config,
