+++ services/worker/tests/job_runners/split/test_first_rows_from_parquet.py
@@ -53,7 +53,7 @@ def get_job_runner(
         )
 
         upsert_response(
-            kind="config-split-names-from-streaming",
+            kind="config-split-names",
             dataset=dataset,
             dataset_git_revision=REVISION_NAME,
             config=config,
