+++ services/worker/tests/job_runners/split/test_first_rows_from_streaming.py
@@ -51,7 +51,7 @@ def get_job_runner(
         )
 
         upsert_response(
-            kind="config-split-names-from-streaming",
+            kind="config-split-names",
             dataset=dataset,
             dataset_git_revision=REVISION_NAME,
             config=config,
