+++ services/worker/tests/job_runners/split/test_opt_in_out_urls_count.py
@@ -47,7 +47,7 @@ def get_job_runner(
         )
 
         upsert_response(
-            kind="config-split-names-from-streaming",
+            kind="config-split-names",
             dataset=dataset,
             dataset_git_revision=REVISION_NAME,
             config=config,
