+++ services/worker/tests/job_runners/split/test_opt_in_out_urls_scan_from_streaming.py
@@ -58,7 +58,7 @@ def get_job_runner(
         )
 
         upsert_response(
-            kind="config-split-names-from-streaming",
+            kind="config-split-names",
             dataset=dataset,
             dataset_git_revision=REVISION_NAME,
             config=config,
