+++ services/worker/tests/job_runners/split/test_split_job_runner.py
@@ -85,7 +85,7 @@ def test_creation(
     )
 
     upsert_response(
-        kind="config-split-names-from-streaming",
+        kind="config-split-names",
         dataset=dataset,
         dataset_git_revision=REVISION_NAME,
         config=config,
