+++ services/worker/tests/test_job_runner_factory.py
@@ -89,7 +89,7 @@ def test_create_job_runner(
 
     if level == "split":
         upsert_response(
-            kind="config-split-names-from-streaming",
+            kind="config-split-names",
             dataset=dataset,
             dataset_git_revision=REVISION_NAME,
             config=config,
